Fixes GS1-128 barcodes without type given
Fixes the the GS1-128 barcode generate for the case when no type is given:

Due to the fact, that the ```GS1128#initialize``` method was not giving
the correct data to the initialize and that the overwritten
```GS1128#data``` returned a different string then what was given in the
```initialize``` method, the ```valid?``` method returned false.

This commit adjusts the initializer so it builds the correct data string
(including the ```FNC1``` symbol at the application identifier)
and removes the data method from the GS1128 class.