Make it possible for PNG barcodes to have transparent background
Currently if you have a barcode in PNG, it will always have white
background. Given PNG supports alpha channel in images, it would be
really nice to be able to use that. This commit adds an option to
specify if you want the background to be transparent but keeps the WHITE
as default.