RMagick outputter should use line command for lines
Newer versions of ImageMagick do not support creating
rectangles that are only a single pixel wide.
https://www.imagemagick.org/discourse-server/viewtopic.php?f=3&p=166063

When drawing a single pixel wide (or tall) line, use
the `#line` command instead of `rectangle`.

This works on older version of ImageMagick like 6.9.6-2
(which supported single pixel wide rectangles), as well
as newer versions like 6.9.10-6 and 7.0.8 (which do NOT
support single pixel wide rectangles).